junit5_app/src/main/java/org/andemar/junit5_app/ejemplos/Account.java
package org.andemar.junit5_app.ejemplos;

import java.math.BigDecimal;
import java.util.Objects;

public class Account {
    private String persona;
    private BigDecimal saldo;

    public Account(String persona, BigDecimal saldo) {
        this.persona = persona;
        this.saldo = saldo;
    }

    public String getPersona() {
        return persona;
    }

    public void setPersona(String persona) {
        this.persona = persona;
    }

    public BigDecimal getSaldo() {
        return saldo;
    }

    public void setSaldo(BigDecimal saldo) {
        this.saldo = saldo;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Account account = (Account) o;
        return Objects.equals(persona, account.persona) && Objects.equals(saldo, account.saldo);
    }

    @Override
    public int hashCode() {
        return Objects.hash(persona, saldo);
    }
}
